django/requirements.txt
django=4.2.7
djangorestframework=3.14.0
pyyaml=6.0.1
uritemplate=4.1.1